Two sets of minor changes:
1. Integrated the EPOC port. All these changes are #ifdefs on the symbol
__SYMBIAN32__. Mostly mucking around with headers, plus EPOC-specific line
editing from Glenn Strong (Glenn.Strong@cs.tcd.ie) and a new command, Pwd,
that prints the cwd.

2. Added an option to compile without Hugs extensions, HASKELL_98_ONLY.
Again, all the changes are #ifdefs on this symbol. The haskell98 flag is
omitted, so all the code which tests it needs to be commented out or
modified, and the toggle struct has the h98 field removed.

Neither set of changes should affect Hugs unless one of these symbols is
defined.