Added need to run convert_hslibs to Readme
Tweaked error cases to reflect the new reality.

There's several static checks still failing.
They're all of the form:

  import Prelude(foo)
  foo = 42

This is legal Haskell code because no-one refers to foo but adding a
reference to foo should elicit an error message and doesn't seem to.
I haven't explored this any further because Sigbjorn's been in this
recently so he'll understand the letter of the law, etc. better than I
do.  (What I think should be done is duplicate the checks and make one
of them a correct name clash (no reference) and one an incorrect class
(ambiguity does arise).)