Fixed dynamic loading implementation for Darwin.
Tweaked the code to use a fresh namespace for each loaded module
instead of the global namespace.  This means that when you lookup
_initModule you get a different answer each time.  What was happening
before was you always got the first _initModule loaded into the
namespace which was always the one from StdDIS.so so we kept initializing
StdDIS instead of initializing the module we'd just loaded.

Someone who knows Darwin better should check this code over.  My
source of documentation was some random web page instead of a proper
manual.  My testcase was an X program being run on a machine on which
I couldn't run X - so all I know is that it loads successfully but I
can't say for sure that it runs ok.