Make the set of suffixes for Haskell modules user-configurable via a -S option, analogous to -P for search paths.  For example, with the invocation
	hugs -S:.y:.ly

the suffix list becomes .hs:.lhs:.y:.ly and when importing a module Foo
it tries Foo.hs, Foo.lhs, Foo.y and Foo.ly.  It is useful in conjunction
with the preprocessor option: the preprocessor can use the suffix to
decide how to turn the file into Haskell.  Perhaps there are other uses.