Tidy up how the interpreter binds the Trex-specific names that are needed during code generation. This is done along the lines of how mdo-related names are handled.
One upshot of this change is that the lifetime of a module's
effective import lists is extended until the user causes
a switch to a different one. This is in general a change
for the better, and it avoids confusing a user that imports
'Trex' qualified & then tries to use records from the
read-eval-print loop.