Murphy's Law caught us again, here's the full story:
There are some broken sed implementations out there which have a fixed
limit on the size of sed scripts, so autoconf tries to be clever and
splits sed scripts into chunks (currently 50 lines). If the value of a
substituted variable spans multiple lines, there is a small chance
that this value gets split into two parts.    :-P   This is a bug in
autoconf, which seems to be known but hard to fix. Thanks to Murphy,
the recent innocent changes triggered this faulty behaviour.

The only safe way to avoid surprises like this in the future is to use
single-line substitutions only.