Use an empty nonterminal rather than error for inserting an initial virtual { in the module, to fix a bug reported by Iavor.
The old code was accepting a module with no module header and an
incomplete last decl like

	f (

as a legal parse (because "topDecls; VARID (" -> error -> "{"), leading
to an internal error back in parseInput.

Unfortunately the same error still occurs with things like

	module Bug where
	f x = x where g (

due to a similar problem with the "end" nonterminal, but I don't know
how to fix that.