Fix for bug reported by Henk-Jan van Tuyl, in which the following corrupts the heap if "foo" isn't found:
	import System.IO

	main :: IO ()
	main = do
		h1 <- openFile "/usr/dict/words" ReadMode
		let fileName = "foo"
		h2 <- openFile fileName ReadMode
		return ()

The second openFile triggers a GC, which corrupts the indirection to
fileName used by the error report.

The bug has been there a long time, but has emerged now because we cut
the starting size of the handle table (to make rare bugs more common).

Today I learned that

	Cell x = ...;
	<possible GC>
	... x ...

is unsafe if x could refer to an indirection, even with conservative GC,
even if there are copies of x on the stack.

The fix is ugly (use a Cell * instead) but I couldn't think of anything
better.