from Neil Mitchell: first page of the options dialog, including overlapping instances, here documents and H98 mode.
Changing H98 mode is marked as "requires restart", but in fact:

If you load a non-haskell 98 doc with +98, it fails. If you then go to
options and set -98, then either :l or :r, it succeeds. If you then go
to options and flip back to +98, :r succeeds (wrongly), but :l fails.